# Optimize HexString.TryParseByte